pyproject.toml
[project]
name = "rahma-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "minio>=7.2.20",
    "psycopg2>=2.9.11",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.45",
    "sqlalchemy-utils>=0.42.1",
    "structlog>=25.5.0",
    "aiosmtplib>=5.0.0",
    "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
